DS-6965 Replace preprocess hook by display mode change
The Drupal display modes have the possibility to show fields as URL.
This mode still allows you to select a image mode. Using this method
removes to need to add a custom variable to a template through a
preprocess hook. This method is also more easily customised.

BREAKING CHANGE: Because a value is removed from a template.