DS-6965 Replace _social_landing_page_get_hero_image with pre-calculated image field
This commit uses the newly added field_landing_page_image to provide a
teaser image for landing pages. This improves the view performance by
moving the calculations for which image to display into the saving of
the entity instead of during every display.

The added benefit of using a standard name of field_{node_type}_image
allows code to be simplified by removing special casings that were added
for landing pages where this field was not yet available.